Update patch set 3
Patch Set 3:

-1 Verified

Not working in the US. The phone type changes to CDMA and goes bogus if GSM mode (Only) is set (reports CDMA phone type, SIM state unknown, and doesn't catch signal). The problem is that it fails to unset CDMA mode. This means everyone going on Roaming to US will get non-working device.

Patch-set: 3