Update patch set 1
Patch Set 1: Verified+1

Reopening the world phone (doesn't work in the US), as it was merged prematurately.

My understanding of the world phone as I viewed it is to support GSM roaming (or using other SIM card) when you're outside US and on a CDMA carrier. But that's not what we want, we want: xt897 - no CDMA wanted, xt897c - world phone.

Technically, the problem is that once the phone detects Sprint signal, it will keep using CDMA SST even if told to be in GSM mode (not relevant to the modem auto switch, which has to be disabled in QPST).

Patch-set: 1
Reviewer: Martin Meloun <6186@1ec6b3db-b2c1-4fa4-84a2-4c7efe89ba71>
Label: Verified=+1