Update patch set 1
Patch Set 1:

While my english is far from perfect, "in the beginning of" sounds wrong to me.

I feel that the correct wording should be 'at the beginning', as 'in the beginning' means rather 'at first', like in 'in the beginning, god created the heaven and the earth'. And no, I'm not feeling religious at all ;).

Other than that, I indeed welcome this notice for translators.

Patch-set: 1